requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
chromadb
python-dotenv
pypdf
google-generativeai